ui: improve responsiveness   components changed: Footer, ProductCard,   EditOrderPage, ProfilePage, AdminProductItems,   CheckoutPage, ShoppingCart
  also renamed: product pages, PurchasePage
