+++ src/components/product/AdminProductItem.jsx
@@ -1,23 +1,32 @@
 import {
-  Button,
-  Flex,
+  Center,
+  Grid,
+  GridItem,
   Image,
   LinkBox,
   LinkOverlay,
-  Td,
   Text,
-  Tr,
 } from "@chakra-ui/react";
 import { Link } from "react-router-dom";
-import DeleteProductAlert from "../info/DeleteProductAlert";
+import DeleteProductAlert from "../modals/DeleteProductAlert";
 
 export default function AdminProductItem({ state }) {
   const { index, item } = state;
 
   return (
-    <Tr data-testid={"product-row"} key={index}>
-      <Td>
-        <Flex>
+    <GridItem data-testid={"product-row"} key={index}>
+      <Grid
+        templateColumns={[null, null, "repeat(5, 1fr)"]}
+        borderBottom={"1px solid black"}
+        borderTop={"1px solid black"}
+        w={"80vw"}
+        my={1}
+      >
+        <GridItem
+          w={["80vw", "80vw", `${80 / 5}vw`]}
+          borderEnd={"1px solid black"}
+          borderStart={"1px solid black"}
+        >
           <Text me={2}>{item.product_name}</Text>
           <LinkBox>
             <LinkOverlay
@@ -30,19 +39,41 @@ export default function AdminProductItem({ state }) {
               <Text>(Edit)</Text>
             </LinkOverlay>
           </LinkBox>
-        </Flex>
-        <Flex>
           <DeleteProductAlert productId={item._id} />
-        </Flex>
-      </Td>
-      <Td>{item.product_price}</Td>
-      <Td>{item.product_description}</Td>
-      <Td>{<Image src={item.product_image} />}</Td>
-      <Td>
-        {item.categories.map((cat, index) => {
-          return <Text key={index + cat}>{cat.category_name}</Text>;
-        })}
-      </Td>
-    </Tr>
+        </GridItem>
+        <GridItem
+          w={["80vw", "80vw", `${80 / 5}vw`]}
+          borderStart={"1px solid black"}
+          borderEnd={"1px solid black"}
+        >
+          {item.product_price}
+        </GridItem>
+        <GridItem
+          w={["80vw", "80vw", `${80 / 5}vw`]}
+          borderStart={"1px solid black"}
+          borderEnd={"1px solid black"}
+        >
+          {item.product_description}
+        </GridItem>
+        <GridItem
+          w={["80vw", "80vw", `${80 / 5}vw`]}
+          borderStart={"1px solid black"}
+          borderEnd={"1px solid black"}
+        >
+          <Center>
+            {<Image w={["60%", "60%", "100%"]} src={item.product_image} />}
+          </Center>
+        </GridItem>
+        <GridItem
+          w={["80vw", "80vw", `${80 / 5}vw`]}
+          borderStart={"1px solid black"}
+          borderEnd={"1px solid black"}
+        >
+          {item.categories.map((cat, index) => {
+            return <Text key={index + cat}>{cat.category_name}</Text>;
+          })}
+        </GridItem>
+      </Grid>
+    </GridItem>
   );
 }
