+++ src/components/product/CartEntry.jsx
@@ -1,4 +1,11 @@
-import { Button, CloseButton, Image, Td, Tr } from "@chakra-ui/react";
+import {
+  Button,
+  CloseButton,
+  Flex,
+  Grid,
+  GridItem,
+  Image,
+} from "@chakra-ui/react";
 import { useContext } from "react";
 import { CartContext } from "../../context/CartContext";
 
@@ -7,43 +14,48 @@ export default function CartEntry(props) {
   const { product } = props;
 
   return (
-    <Tr h={20}>
-      <Td px={1} textAlign={"center"}>
+    <Grid h={20} templateColumns={"repeat(4, 1fr)"}>
+      <GridItem px={1} textAlign={"center"}>
         {product.product_name}
-      </Td>
-      <Td px={1} textAlign={"center"}>{`$${product.product_price}`}</Td>
-      <Td px={1} textAlign={"center"}>
-        <Image src={product.product_image} />
-      </Td>
-      <Td px={1} textAlign={"center"}>
-        <Button
-          variant={"ghost"}
-          h={"1em"}
-          onClick={() => {
-            removeFromCart(product);
-          }}
-        >
-          -
-        </Button>
-        {product.quantity}
-        <Button
-          variant={"ghost"}
-          h={"1em"}
-          onClick={() => {
-            addToCart(product);
-          }}
-        >
-          +
-        </Button>
-      </Td>
-
+      </GridItem>
+      <GridItem
+        px={1}
+        textAlign={"center"}
+      >{`$${product.product_price}`}</GridItem>
+      <GridItem px={1} textAlign={"center"}>
+        <Image src={product.product_image} h={"70px"} w={"70px"} />
+      </GridItem>
+      <GridItem px={1} textAlign={"center"}>
+        <Flex>
+          <Button
+            variant={"ghost"}
+            h={"1em"}
+            onClick={() => {
+              removeFromCart(product);
+            }}
+          >
+            -
+          </Button>
+          {product.quantity}
+          <Button
+            variant={"ghost"}
+            h={"1em"}
+            onClick={() => {
+              addToCart(product);
+            }}
+          >
+            +
+          </Button>
+        </Flex>
+      </GridItem>
       <CloseButton
         size={"sm"}
         position={"relative"}
-        right={7}
-        bottom={-1}
+        right={"20px"}
+        bottom={"70"}
+        border={"1px solid black"}
         onClick={() => removeFromCart(product, true)}
       />
-    </Tr>
+    </Grid>
   );
 }
